Update ES query, bring in bluebird.
Update ES plain terms query with three levels of nested aggregation for
aggregating the sum of counts within a year and then within a ngram
phrase across the years. Also populate 0 if there is no data available
for a particular year.
Also the total counts should contain all 1 to 6 ngrams counts across all
of the corpus and is not per year.

Add bluebird dependency to just utilize Promise.map concurrency feature
so that the requests to ES for indexing is throttled.